refactor: 명단 삭제 기능 response 메시지 추가

+++ backend/Group3/src/main/java/SW_Engineering/Group3/service/ClubService.java
@@ -273,7 +273,7 @@ public class ClubService {
      */
 
     @Transactional
-    public void deleteMember(String targetStudentId, Long clubId) {
+    public boolean deleteMember(String targetStudentId, Long clubId) {
         Member targetMember = memberRepository.findByStudentId(targetStudentId).orElseGet(null);
 
         ClubMemberList targetClubMember = clubMemberRepository.findByMember(targetMember).stream()
@@ -281,7 +281,10 @@ public class ClubService {
                 .findAny()
                 .orElse(null);
 
+        if (targetMember == null || targetClubMember == null) return false;
+
         clubMemberRepository.delete(targetClubMember);
+        return true;
     }
 
     public List<Club> getAllClubs() {
